opt: move `buildUDF` from optbuilder/scalar.go to optbuilder/routine.go
This commit moves logic that will be shared by both UDFs and procedures
into the same file.

Release note: None
